refactor: deleted unused files and universal stylesheet
