+++ native-client/App.js
@@ -1,14 +1,9 @@
 import React from 'react'
-import { Alert, Platform, StatusBar, StyleSheet, View } from 'react-native'
+import { Platform, StyleSheet, View } from 'react-native'
 import RootNavigation from './src/navigation/RootNavigation'
 import { Font, AppLoading } from 'expo'
+import styles from './assets/stylesheet'
 
-let API_ROOT
-// if (IS_SIM) {
-//   API_ROOT = 'http://localhost:5000'
-// } else {
-API_ROOT = 'http://192.168.1.190:5000'
-// }
 
 export default class App extends React.Component {
 
@@ -33,15 +28,6 @@ export default class App extends React.Component {
         }
     // async function _loadAssetsAsync() {
 
-
-
-    _onPressButton() {
-        fetch(`${API_ROOT}/users`)
-            .then((res) => res.text())
-            .then((data) => Alert.alert(data))
-        // Alert.alert('You tapped the button!');
-    }
-
     render() {
         if (!this.state.fontLoaded) {
             return (
@@ -55,29 +41,12 @@ export default class App extends React.Component {
 
       return (
       <View style={styles.container}>
-        {/* <Text style={styles.title}>fetch</Text>
-        <TouchableHighlight onPress={this._onPressButton} underlayColor="white">
-          <View>
-            <Text>get started</Text>
-          </View>
-        </TouchableHighlight> */}
-        <View style={{height:20}}>
+        {/* what is this element? */}
+        {/* <View>
          {Platform.OS === 'ios' && <StatusBar barStyle="light-content" />}
           {Platform.OS === 'android' && <View style={styles.statusBarUnderlay} />}
-        </View>
+        </View> */}
         <RootNavigation />
       </View>
     )}
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: 'black',
-  },
-  title: {
-    fontFamily: 'Arial',
-    fontSize: 70,
-    color: 'black'
-  }
-})
