+++ native-client/assets/stylesheet.js
@@ -1,0 +1,25 @@
+import { StyleSheet } from 'react-native'
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: 'black',
+  },
+  title: {
+    fontFamily: 'Arial',
+    fontSize: 70,
+    color: 'white'
+  },
+  resultsContainer: {
+    flex: 1,
+    paddingTop: 15,
+    backgroundColor: 'black',
+  },
+  text: {
+    fontFamily: 'Arial',
+    fontSize: 20,
+    color: 'white'
+  }
+})
+
+export default styles
